requirements.txt
biopython==1.79
joblib==1.0.1
numpy==1.20.3
scikit-learn==1.0.2
scipy==1.7.3
threadpoolctl==2.1.0
torch==1.8.1
typing-extensions==3.10.0.0
